добавлена проверка на права пользователя при работе с серверами и наборами, выборка стиля из уже существующих

+++ project/WMSMash/tree/views.py
@@ -58,6 +58,9 @@ def register(request):
 def show_category_tree(request):
   cursor = connection.cursor()
 
+  # User running the application
+  USER = 1
+
   # Definition of the seected server and set of layers
   slt_server = request.GET.get('list_servers', 0)
   slt_set = request.GET.get('list_sets', 0)
@@ -70,15 +73,26 @@ def show_category_tree(request):
   
   # Saved a list of layers which couldn't be inserted into a set of layers
   errlayer = "" 
-  
+
+  # Stores error messages
+  errors = ""
+    
   # Operations with layers of virtual set of layers
   oprt = request.GET.get('oprt', 0)
   # Layer in set that is processed
   set_layer = request.GET.get('set_layer', -1)
-  
-  if oprt and (set_layer != -1):
+  set_obj = 0
+  if slt_set:
+    try:
+      set_obj = LayerSet.objects.get(id = slt_set)
+    except:
+      set_obj = 0
+      slt_set = 0
+      errors = "Such a set was not created"
+  if oprt and (set_layer != -1) and set_obj:
+
     # Operation of adding a layers in a virtual set of layers
-    if oprt == 'add':
+    if ((oprt == 'add') and (set_obj.author_id == USER or set_obj.pub == 1)):
       layer = request.GET.get('layer', -1)
       namelayer = request.GET.get('namelayer', 0)
       list_layers = request.GET.get('layers', -1)
@@ -90,11 +104,11 @@ def show_category_tree(request):
           errlayer += "%s ?end %d %d;"%(i[0], i[1], i[2])
     
     # Operation of deleting a layers in a virtual set of layers
-    if oprt == 'del':
+    if ((oprt == 'del') and (set_obj.author_id == USER or set_obj.pub == 1)):
       LayerTree.objects.del_fromset(set_layer, cursor)
     
     # Operation of off hidden layers
-    if (oprt == 'hidd_off') and set_layer:
+    if ((oprt == 'hidd_off') and set_layer and (set_obj.author_id == USER or set_obj.pub == 1)):
       layers = []
       if set_layer == -1:
         ll = (0, )
@@ -104,7 +118,7 @@ def show_category_tree(request):
       LayerTree.objects.hidden(1, layers, cursor)
     
     # Operation of on hidden layers 
-    if (oprt == 'hidd_on') and set_layer:
+    if ((oprt == 'hidd_on') and set_layer and (set_obj.author_id == USER or set_obj.pub == 1)):
       layers = []
       if set_layer == -1:
         ll = (0, )
@@ -114,17 +128,17 @@ def show_category_tree(request):
       LayerTree.objects.hidden(0, layers, cursor)
     
     # Operation of adding style to a layer of a set
-    if oprt == 'add_style':
+    if ((oprt == 'add_style') and (set_obj.author_id == USER or set_obj.pub == 1)):
       name = request.GET.get('namelayer', 0)
       url = request.GET.get('sld', 0)
       LayerTree.objects.addstyle(set_layer, name, url)
     
     # Operation of deleting style to a layer of a set
-    if oprt == 'del_style':
+    if ((oprt == 'del_style') and (set_obj.author_id == USER or set_obj.pub == 1)):
       LayerTree.objects.delstyle(set_layer)
     
     # Operation off the layer of publicity
-    #if oprt == 'pub_off':
+    #if ((oprt == 'pub_off') and (set_obj.author_id == USER or set_obj.pub == 1)):
     #  layers = []
     #  if set_layer == -1:
     #     ll = (0, )
@@ -134,16 +148,16 @@ def show_category_tree(request):
     #  LayerTree.objects.hidden(0, layers, cursor)
 
   else:
-    # Operation of adding a new group in set of layers 
-    title_group = request.GET.get('title_group', 0)
-    if title_group:
-      lparent = request.GET.get('where_layer', -1)
-      list_set = request.GET.get('list_sets', 0)
-      namegroup = LayerTree.objects.add_newgroup(title_group, lparent, list_set)
+    if set_obj:
+      # Operation of adding a new group in set of layers 
+      title_group = request.GET.get('title_group', 0)
+      if (title_group and (set_obj.author_id == USER or set_obj.pub == 1)):
+        lparent = request.GET.get('where_layer', -1)
+        list_set = request.GET.get('list_sets', 0)
+        namegroup = LayerTree.objects.add_newgroup(title_group, lparent, list_set)
 
     # Operations managment sets
   
-    susers = 1 #hz
     # Operation to create a new set of layers
     addset = request.GET.get('add_set', 0)
     if addset:
@@ -151,25 +165,31 @@ def show_category_tree(request):
       stitle = request.GET.get('title_set', 0)
       sabstr = request.GET.get('abstract_set', 0)
       skeywords = request.GET.get('keywords_set', 0)
-
-      slt_set = (LayerSet.objects.add_newset(sname, stitle, sabstr, skeywords, susers)).id
+      spub = int(request.GET.get('pub_set', 0))
+      slt_set = (LayerSet.objects.add_newset(sname, stitle, sabstr, skeywords, int(spub), USER)).id
 
     # Operation of editing a describe set of layers
     editset = request.GET.get('edit_set', 0)
-    if editset:
+    if editset and set_obj:
       sname = request.GET.get('name_set', 0)
       stitle = request.GET.get('title_set', 0)
       sabstr = request.GET.get('abstract_set', 0)
       skeywords = request.GET.get('keywords_set', 0)
-
-      LayerSet.objects.edit_set(slt_set, sname, stitle, sabstr, skeywords, susers)
+      spub = request.GET.get('pub_set', 0)
+      if set_obj.author_id == USER:
+        LayerSet.objects.edit_set(slt_set, sname, stitle, sabstr, skeywords, int(spub), USER)
+      else:
+        errors = "This set can edit only the owner"
   
     # Operation of remove a set of layers
     delset = request.GET.get('delete_set', 0)
-    if delset:
-      LayerSet.objects.del_set(slt_set, cursor)
-      slt_set = 0
-  
+    if delset and set_obj:
+      if set_obj.author_id == USER:
+        LayerSet.objects.del_set(slt_set, cursor)
+        slt_set = 0
+      else:
+        errors = "This set can be removed only by the owner"
+
     # Operations managment servers
   
     # Operation to add, edit the URL and update server
@@ -177,20 +197,50 @@ def show_category_tree(request):
     if servers:
       if servers.find('add_') == 0:
         Servers.objects.add(servers)
+      else:
+        try:
+          serv_obj = Servers.objects.get(id = slt_server)
+        except:
+          serv_obj = 0
+          slt_server = 0
+          errors = "Such a server was not added"
       if servers.find('edit_') == 0:
-        Servers.objects.editURL(slt_server, servers)
+        if serv_obj:
+          if serv_obj.owner_id == USER:
+            Servers.objects.editURL(slt_server, servers)
+          else:
+            errors = "URL of this server can edit only the resource owner"
       if servers == 'update':
-        Servers.objects.update(slt_server)
+        if serv_obj:
+          if serv_obj.owner_id == USER:
+            Servers.objects.update(slt_server)
+          else:
+            errors = "URL of this server can update only the resource owner"
       
     # Operation of remove server
     delserv = request.GET.get('delete_server', 0)
     if delserv:
-      name_sets = Servers.objects.delete(slt_server, cursor)
-      if not name_sets:
+      try:
+        serv_obj = Servers.objects.get(id = slt_server)
+      except:
+        serv_obj = 0
         slt_server = 0
+        errors = "Such a server was not added"
+      if serv_obj:
+        if serv_obj.owner_id == USER:
+          name_sets = Servers.objects.delete(slt_server, cursor)
+          if not name_sets:
+            slt_server = 0
+        else:
+          errors = "This server can only delete the resource owner"
 
   # Formed a list of servers with class = "select_style", value = Servers.id, id = "list_servers"
-  BIG_CHOICES = [(c.id, "(%s(%s"%(c.title, c.name)) for c in Servers.objects.all()]
+  #BIG_CHOICES = [(c.id, "(%s(%s"%(c.title, c.name)) for c in Servers.objects.all()]
+  BIG_CHOICES = []
+  for c in Servers.objects.all():
+    if (c.owner_id == USER) or (c.pub == 1):
+      BIG_CHOICES.append((c.id, "(%s(%s" % (c.title, c.name)))
+
   list_servers = forms.ChoiceField(choices = BIG_CHOICES,
                                      required = False,
                                      label = '',
@@ -200,7 +250,10 @@ def show_category_tree(request):
                                     ) 
   
   # Formed a list of sets with class = "select_style", value = LayerSet.id, id = "list_sets"
-  BIG_CHOICES = [(c.id, "(%s(%s"%(c.title, c.name)) for c in LayerSet.objects.all()]
+  BIG_CHOICES = []
+  for c in LayerSet.objects.all():
+    if (c.author_id == USER) or (c.pub == 1):
+      BIG_CHOICES.append((c.id, "(%s(%s"%(c.title, c.name)))
   list_sets = forms.ChoiceField(choices = BIG_CHOICES,
                                   required = False,
                                   label = '',
@@ -210,7 +263,12 @@ def show_category_tree(request):
                                  )
   
   # Formed a list of SLD with class = "select_small", value = SLD.id, id = "list_sld"
-  BIG_CHOICES = [(c.id, "%s(%s)"%(c.name, c.url)) for c in SLD.objects.all()]
+  #BIG_CHOICES = [(c.id, "%s(%s)"%(c.name, c.url)) for c in SLD.objects.all()]
+  BIG_CHOICES = []
+  for c in SLD.objects.all():
+    if (c.owner_id == USER or not c.owner_id):
+      BIG_CHOICES.append((c.name+c.url, "%s(%s)"%(c.name, c.url)))
+  
   list_sld = forms.ChoiceField(choices = BIG_CHOICES,
                                      required = False,
                                      label = '',
@@ -227,8 +285,9 @@ def show_category_tree(request):
   if not slt_server:
     # Take the first server from the list
     for c in Servers.objects.all():
-      selected_server = c
-      break;
+      if (c.pub == 1) or (c.owner_id == USER):
+        selected_server = c
+        break;
     
     if not selected_server:
       # If list is empty then id is not defined
@@ -246,8 +305,9 @@ def show_category_tree(request):
   if not slt_set:
     # Take the first set from the list
     for c in LayerSet.objects.all():
-      selected_set = c
-      break;
+      if (c.pub == 1) or (c.author_id == USER):
+        selected_set = c
+        break;
     
     if not selected_set:
       # If list is empty then id is not defined
@@ -266,19 +326,22 @@ def show_category_tree(request):
                       AS 
                       (
                         SELECT 
-                              name, title, id, parent_id, server_id, NULL::varchar AS parent_name, id::text AS path 
+                              name, title, id, parent_id, server_id, pub, NULL::varchar AS parent_name, id::text AS path 
                         FROM 
                               tree_layers
                         WHERE parent_id IS NULL
                         UNION 
                         SELECT 
-                              f1.name, f1.title, f1.id, f1.parent_id, f1.server_id, tree.name 
+                              f1.name, f1.title, f1.id, f1.parent_id, f1.server_id, f1.pub, tree.name 
                         AS parent_name, tree.path || '-000-'||f1.id::text AS path 
                         FROM 
                               tree 
                         JOIN tree_layers f1 ON f1.parent_id = tree.id)
-                        SELECT name, title, id, parent_id, name, path FROM tree WHERE server_id = %s ORDER BY path, title;
-                   ''', (id_server, ))
+                        SELECT tree.name, tree.title, tree.id, tree.parent_id, tree.pub, tree.name, tree.path 
+                        FROM tree, tree_servers 
+                        WHERE tree.server_id = %s and (tree_servers.owner_id = %s or tree.pub = true) and tree_servers.id = tree.server_id
+                        ORDER BY tree.path, tree.title;
+                   ''', (id_server, USER))
     layers = cursor.fetchall()
   else:
     layers = None
@@ -301,8 +364,18 @@ def show_category_tree(request):
                         FROM 
                               tree 
                         JOIN tree_layertree f1 ON f1.parent_id = tree.id)
-                        SELECT name, id, parent_id, lset_id, layer_id, hidden, sld_id, name, path FROM tree WHERE lset_id = %s ORDER BY path, name;
-                  ''', (id_set, ))
+                        SELECT tree.name, tree.id, tree.parent_id, tree.lset_id, tree.layer_id, tree.hidden, tree.sld_id, tree.name, tree.path 
+                        FROM tree, tree_layers, tree_servers
+                        WHERE tree.lset_id = %s AND 
+                              (
+                                (tree.layer_id != NULL and tree.layer_id = tree_layers.id AND tree_layers.server_id = tree_servers.id
+                                and tree_servers.owner_id = %s or tree_layers.pub = true)
+                              OR
+                                tree.layer_id IS NULL
+                              )
+                        GROUP BY tree.name, tree.id, tree.parent_id, tree.lset_id, tree.layer_id, tree.hidden, tree.sld_id, tree.path
+                        ORDER BY tree.path, tree.name;
+                  ''', (id_set, USER))
     layertree = cursor.fetchall()
   else:
     layertree = None
@@ -314,6 +387,7 @@ def show_category_tree(request):
                                 'layers' : layers,
                                 'nodes_layertree':LayerTree.tree.all(), 
                                 'layertree' : layertree,
+                                'sld_objects': SLD.objects.all(),
                                 'list_servers': list_servers.widget.render("list_servers", id_server),
                                 'list_sets': list_sets.widget.render("list_sets", id_set),
                                 'list_sld': list_sld.widget.render("list_sld", None),
@@ -321,7 +395,8 @@ def show_category_tree(request):
                                 'selected_set':selected_set,
                                 'name_sets': name_sets,
                                 'error_layers': errlayer,
-                                'namegroup': namegroup
+                                'namegroup': namegroup,
+                                'errors': errors
                               },
                               context_instance=RequestContext(request)
                             )
